+++ yash-builtin/src/kill/signal.rs
@@ -24,7 +24,7 @@ use std::fmt::Display;
 use std::str::FromStr;
 use yash_env::semantics::ExitStatus;
 use yash_env::signal::{Name, Number, RawNumber};
-use yash_env::system::{System, SystemEx};
+use yash_env::system::System;
 
 /// Specification of a signal to be sent by the kill built-in
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
@@ -104,7 +104,7 @@ impl Signal {
     /// as an exit status and converted to a signal number. The function returns
     /// `None` if `self` does not represent a valid signal.
     #[must_use]
-    pub fn to_name_and_number<S: SystemEx>(self, system: &S) -> Option<(Name, Number)> {
+    pub fn to_name_and_number<S: System>(self, system: &S) -> Option<(Name, Number)> {
         match self {
             Signal::Name(name) => Some((name, system.signal_number_from_name(name)?)),
             Signal::Number(number) => {